feat: Add voucher application functionality and related DTOs
- Implemented applyVoucher method in VoucherService to handle voucher application logic.
- Created ApplyVoucherRequest and ApplyVoucherResponse DTOs for voucher application.
- Updated VoucherController to include endpoint for applying vouchers.

+++ src/main/java/com/thanh/foodOrder/feature/voucher/domain/Voucher.java
@@ -1,8 +1,8 @@
 package com.thanh.foodorder.feature.voucher.domain;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
-
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -29,7 +29,7 @@ public class Voucher {
     private Long id;
     private String code;
     private int percentDiscount;
-    private int maxDiscount;
+    private BigDecimal maxDiscount;
 
     private LocalDate expiration;
     private String createdBy;
